+++ NEWS.md
@@ -1,3 +1,11 @@
+# rlandfire 0.4.0
+  - Added new function (`getZone()`) which returns the LANDFIRE map zone based on a spatial object or zone name 
+  - Removed dependencies on `terra` and `stringr`
+  - Updated to MIT + file LICENSE
+  - Added argument to specify order of values in the `extend` argument of `getAOI()`
+  - Added additional tests for `getAOI()`, `getZone()`, and `landfireAPI()`
+  - Improved portability of examples
+
 # rlandfire 0.3.0
 
 -   Added new function (`viewProducts()`) to open the LFPS products table in your web browser